Adding lazy logging along critical path
`f-strings` are evaluated before passing them down to the logging
functions (even if they are not used). By using boring (and IMO
harder to read) `%s` or `%r` we can delay or elude that work
depending on the logging settings.

With debug level logging disabled, this won't necessarily save a
noticable amount of run time on one individual run but will
save CPU cycles none the less.